remove icons
