+++ trilogy/__init__.py
@@ -4,6 +4,6 @@ from trilogy.dialect.enums import Dialects
 from trilogy.executor import Executor
 from trilogy.parser import parse
 
-__version__ = "0.0.3.95"
+__version__ = "0.0.3.96"
 
 __all__ = ["parse", "Executor", "Dialects", "Environment", "CONFIG"]
