+++ trilogy/authoring/__init__.py
@@ -60,47 +60,16 @@ from trilogy.core.statements.author import (
     RowsetDerivationStatement,
     SelectItem,
     SelectStatement,
+    ShowCategory,
+    ShowStatement,
+    ValidateStatement,
 )
 from trilogy.parsing.common import arbitrary_to_concept, arg_to_datatype
 
 __all__ = [
-    "Concept",
-    "Function",
-    "WhereClause",
-    "Comparison",
-    "FilterItem",
-    "CaseWhen",
-    "CaseElse",
-    "AggregateWrapper",
-    "WindowItem",
-    "WindowOrder",
-    "WindowType",
-    "WindowItemOrder",
-    "WindowItemOver",
-    "DataType",
-    "StructType",
-    "ArrayType",
-    "NumericType",
-    "Grain",
-    "RowsetDerivationStatement",
-    "MapType",
-    "ListWrapper",
-    "FunctionType",
-    "FunctionFactory",
-    "ConceptDeclarationStatement",
-    "ConceptTransform",
-    "SelectItem",
-    "SelectStatement",
-    "Environment",
-    "ConceptRef",
-    "HavingClause",
-    "MagicConstants",
-    "Metadata",
-    "OrderBy",
-    "OrderItem",
-    "Parenthetical",
-    "SubselectComparison",
-    "Conditional",
+    # trilogy.constants
+    "DEFAULT_NAMESPACE",
+    # trilogy.core.enums
     "BooleanOperator",
     "ComparisonOperator",
     "FunctionClass",
@@ -108,20 +77,65 @@ __all__ = [
     "InfiniteFunctionArgs",
     "Ordering",
     "Purpose",
-    "DEFAULT_NAMESPACE",
-    "arbitrary_to_concept",
-    "arg_to_datatype",
+    # trilogy.core.functions
+    "FunctionFactory",
+    # trilogy.core.models.author
+    "AggregateWrapper",
+    "CaseElse",
+    "CaseWhen",
+    "Comparison",
+    "Concept",
+    "ConceptRef",
+    "Conditional",
+    "FilterItem",
+    "Function",
+    "FunctionCallWrapper",
+    "HavingClause",
+    "MagicConstants",
+    "Metadata",
+    "MultiSelectLineage",
+    "OrderBy",
+    "OrderItem",
+    "Parenthetical",
+    "RowsetItem",
+    "SubselectComparison",
+    "WhereClause",
+    "WindowItem",
+    "WindowItemOrder",
+    "WindowItemOver",
+    "WindowOrder",
+    "WindowType",
+    # trilogy.core.models.core
+    "ArrayType",
+    "DataType",
+    "ListWrapper",
+    "MapType",
+    "NumericType",
+    "StructType",
+    "TraitDataType",
+    # trilogy.core.models.datasource
+    "Address",
+    "Datasource",
+    "DatasourceMetadata",
+    # trilogy.core.models.environment
+    "Environment",
+    # trilogy.core.statements.author
+    "ConceptDeclarationStatement",
+    "ConceptTransform",
+    "CopyStatement",
+    "Grain",
+    "HasUUID",
+    "ImportStatement",
     "MultiSelectStatement",
     "PersistStatement",
     "RawSQLStatement",
-    "Datasource",
-    "DatasourceMetadata",
-    "MultiSelectLineage",
-    "RowsetItem",
-    "FunctionCallWrapper",
-    "CopyStatement",
-    "HasUUID",
-    "ImportStatement",
-    "Address",
-    "TraitDataType",
+    "RowsetDerivationStatement",
+    "SelectItem",
+    "SelectStatement",
+    "ShowCategory",
+    "ShowStatement",
+    "ValidateStatement",
+    # trilogy.parsing.common
+    "arbitrary_to_concept",
+    "arg_to_datatype",
 ]
